requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
branca==0.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-widget-tweaks==1.4.8
entrypoints==0.3
folium==0.12.1
gTTS==2.2.2
gunicorn==20.1.0
idna==2.10
imutils==0.5.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
many-stop-words==0.2.2
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
pygame==2.0.1
Pygments==2.8.1
pyit2fls==0.6.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pywinpty==0.5.7
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
scikit-fuzzy==0.4.2
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
